analyze/requirements.txt
requests>=2.25.1
pandas>=1.3.0
numpy>=1.21.0
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0
pandas-ta>=0.3.14b0
